requirements.txt

django==4.1
django-dotenv
pudb
django-widget-tweaks
requests
